Branch1

+++ AlutechShopDiploma/Models/Entities/Good.cs
@@ -11,7 +11,7 @@ namespace AlutechShopDiploma.Models.Entities
         private string name;
         private string description;
         private double price;
-        private int rating;
+        private double rating;
         private int views;
         private int tableID;
         private int itemID;
@@ -20,7 +20,7 @@ namespace AlutechShopDiploma.Models.Entities
         public string Name { get => name; set => name = value; }
         public string Description { get => description; set => description = value; }
         public double Price { get => price; set => price = value; }
-        public int Rating { get => rating; set => rating = value; }
+        public double Rating { get => rating; set => rating = value; }
         public int Views { get => views; set => views = value; }
         public int TableID { get => tableID; set => tableID = value; }
         public int ItemID { get => itemID; set => itemID = value; }
